commit 29.12.2024

+++ javaProject/src/main/java/com/java/javaProject/Service/ProductService.java
@@ -1,6 +1,8 @@
 package com.java.javaProject.Service;
 import java.util.Optional;
 import com.java.javaProject.Entity.Product;
+import com.java.javaProject.Entity.Order;
+import com.java.javaProject.Repository.OrderRepository;
 import com.java.javaProject.Repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,8 +10,13 @@ import java.util.List;
 
 @Service
 public class ProductService implements IProductService {
+	
 	@Autowired
 	private ProductRepository productRepository;
+	
+	@Autowired
+	private OrderRepository orderRepository;
+	
 	@Override
 	public List<Product> getAllProducts() {
 	    List<Product> products = productRepository.findAll();
@@ -22,8 +29,15 @@ public class ProductService implements IProductService {
 	}
 
 	@Override
-	public void deleteProduct(Long id) {
-		productRepository.deleteById(id);
+	public String deleteProduct(Long id) {
+        List<Order> orders = orderRepository.findByProductId(id);
+        
+        if (!orders.isEmpty()) {
+            return "Bu ürüne bağlı siparişler bulunduğu için ürün silinemiyor.";
+        }
+
+        productRepository.deleteById(id);
+        return null;
 	}
 	
 	@Override
